Store replicas in the segments
This will allow to get more visibility on which nodes are involved in segments, reduce the number of JMX calls to perform when starting a segment.
It is also a first step to build a new segment scheduler that will not rely on JMX metrics for limiting the number of repair per node but rather use backend storage with locks.